Implement a new version of DBSCAN
Right now, we do two full passes overall all points. The first pass is
counting the number of neighbors of each point, and the second produces
cluster labels.

We don't have to go through all points in the second loop. The points
that are not near any points are not going to be used, as well as the
points that are not near any of the core points. It should be possible
to transform the second loop to go over only the core points,
potentially significantly reducing the size of the loop.