Add a test for a subtle DBSCAN situation
Situation: a core point is connected to only boundary points. However,
those boundary are connected to other core points. So it may happen that
the boundary points are assigned to other clusters, resulting in a
cluster with a single point (core).

This test makes sure that we don't accidentally label it as noise, as
its label is likely to be itself.